+++ test/roda/list_grid_data_test.rb
@@ -12,6 +12,8 @@ BASIC_EXPECTED = [
 
 ALLOW_ACCESS = ->(function, program, permission) { false }
 DENY_ACCESS = ->(function, program, permission) { true }
+TRUE_CLIENT_RULE = ->(args) { true }
+FALSE_CLIENT_RULE = ->(args) { false }
 
 HAS_PERMISSION = ->(args) { true }
 NO_PERMISSION = ->(args) { false }
@@ -174,36 +176,42 @@ class ListGridDataTest < Minitest::Test
     assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(root_path: '/a/b/c') }
     assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid') }
     assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', deny_access: ALLOW_ACCESS) }
-    assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION) }
+    assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE) }
     assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', has_permission: HAS_PERMISSION) }
     assert_raises(KeyError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c') }
   end
 
   def test_params
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: basic_loader)
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: basic_loader)
     assert_nil data.params
 
     params = { one: 1 }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: basic_loader, params: params)
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: basic_loader, params: params)
     assert_equal({ one: 1 }, data.params)
 
     params = { key: 'standard', sub_type_id: '3', product_column_ids: '[73, 74]' }
     additions = { conditions: { standard: [{ col: 'id', op: '=', val: '$:sub_type_id$' }] } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: params)
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: params)
     assert_equal '3', data.params[:sub_type_id]
 
     params = { query_string: 'key=standard&sub_type_id=3&product_column_ids=[73, 74]' }
     additions = { conditions: { standard: [{ col: 'id', op: '=', val: '$:sub_type_id$' }] } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: params)
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: params)
     assert_equal '3', data.params['sub_type_id']
 
     expect = [{ col: 'id', op: '=', val: '3' }]
+    assert_equal expect, data.conditions
+
+    params = { key: 'standard', sub_type_id: nil, product_column_ids: '[73, 74]' }
+    additions = { conditions: { standard: [{ col: 'id', op: '=', val: '$:sub_type_id$', optional: true }] } }
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: params)
+    expect = [{ col: 'id', op: '=', val: '', optional: true }]
     assert_equal expect, data.conditions
   end
 
   def test_basic_grid
     DB.array_expect(BASIC_DATA)
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: basic_loader)
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: basic_loader)
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -229,7 +237,7 @@ class ListGridDataTest < Minitest::Test
   def test_multiselect
     DB.array_expect(BASIC_DATA)
     additions = { multiselect: { multitest: { grid_caption: 'Multi caption', url: '/d/e/f' } } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, multi_key: 'multitest', config_loader: loader_extended(additions))
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, multi_key: 'multitest', config_loader: loader_extended(additions))
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -240,7 +248,7 @@ class ListGridDataTest < Minitest::Test
     assert_equal [], tester['multiselect_ids']
 
     additions = { multiselect: { multitest: { grid_caption: 'Multi caption', url: '/d/e/f', preselect: 'SELECT id FROM users' } } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, multi_key: 'multitest', config_loader: loader_extended(additions), params: { id: 1 })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, multi_key: 'multitest', config_loader: loader_extended(additions), params: { id: 1 })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -252,7 +260,7 @@ class ListGridDataTest < Minitest::Test
   def test_tree
     DB.array_expect(BASIC_DATA)
     additions = { tree: { tree_column: 'tcol', tree_caption: 'The Tree', suppress_node_counts: false, groupDefaultExpanded: -1 } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions))
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions))
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -272,7 +280,7 @@ class ListGridDataTest < Minitest::Test
                     auth: { function: 'security', program: 'menu', permission: 'edit'}},
                    {separator: true},
                    {url: '/development/masterfiles/users/$:id$', text: 'delete', icon: 'delete', is_delete: true, popup: true}] }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: { id: '1' })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -308,7 +316,7 @@ class ListGridDataTest < Minitest::Test
                     auth: { function: 'security', program: 'menu', permission: 'edit'} },
                    {separator: true},
                    {url: '/development/masterfiles/users/$:id$', text: 'delete', icon: 'delete', is_delete: true, popup: true}] }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: DENY_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: { id: '1' })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: DENY_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -328,7 +336,7 @@ class ListGridDataTest < Minitest::Test
                     has_permission: [ :key1, :key2] },
                    {separator: true},
                    {url: '/development/masterfiles/users/$:id$', text: 'delete', icon: 'delete', is_delete: true, popup: true}] }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: { id: '1' })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -348,7 +356,7 @@ class ListGridDataTest < Minitest::Test
                     has_permission: [ :key1, :key2] },
                    {separator: true},
                    {url: '/development/masterfiles/users/$:id$', text: 'delete', icon: 'delete', is_delete: true, popup: true}] }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: NO_PERMISSION, config_loader: loader_extended(additions), params: { id: '1' })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: NO_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -360,17 +368,57 @@ class ListGridDataTest < Minitest::Test
     assert_equal expected, actions
   end
 
+  def test_hide_for_client_rule_hide_actions
+    DB.array_expect(BASIC_DATA)
+    additions = { actions:
+                  [{url: '/development/masterfiles/users/$:id$', text: 'view', icon: 'view-show', title: 'View', popup: true},
+                   {url: '/development/masterfiles/users/$:id$/edit', text: 'edit', icon: 'edit', title: 'Edit',
+                    hide_for_client_rule: 'CR_PROD.some_check?' },
+                   {separator: true},
+                   {url: '/development/masterfiles/users/$:id$', text: 'delete', icon: 'delete', is_delete: true, popup: true}] }
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: TRUE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
+    rows = nil
+    data.stub(:load_report_def, BASIC_DM_REPORT) do
+      rows = data.list_rows
+    end
+    tester = JSON.parse(rows)
+    assert_equal BASIC_EXPECTED, tester['rowDefs']
+    actions = tester['columnDefs'].first['valueGetter']
+    expected = "[{\"text\":\"view\",\"url\":\"/development/masterfiles/users/$col0$\",\"col0\":\"id\",\"icon\":\"view-show\",\"title\":\"View\",\"popup\":true},{\"text\":\"sep01\",\"is_separator\":true},{\"text\":\"delete\",\"url\":\"/development/masterfiles/users/$col0$\",\"col0\":\"id\",\"prompt\":\"Are you sure?\",\"method\":\"delete\",\"icon\":\"delete\",\"popup\":true}]"
+    assert_equal expected, actions
+  end
+
+  def test_hide_for_client_rule_show_actions
+    DB.array_expect(BASIC_DATA)
+    additions = { actions:
+                  [{url: '/development/masterfiles/users/$:id$', text: 'view', icon: 'view-show', title: 'View', popup: true},
+                   {url: '/development/masterfiles/users/$:id$/edit', text: 'edit', icon: 'edit', title: 'Edit',
+                    hide_for_client_rule: 'CR_PROD.some_check?' },
+                   {separator: true},
+                   {url: '/development/masterfiles/users/$:id$', text: 'delete', icon: 'delete', is_delete: true, popup: true}] }
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
+    rows = nil
+    data.stub(:load_report_def, BASIC_DM_REPORT) do
+      rows = data.list_rows
+    end
+    tester = JSON.parse(rows)
+    assert_equal BASIC_EXPECTED, tester['rowDefs']
+    actions = tester['columnDefs'].first['valueGetter']
+    expected = "[{\"text\":\"view\",\"url\":\"/development/masterfiles/users/$col0$\",\"col0\":\"id\",\"icon\":\"view-show\",\"title\":\"View\",\"popup\":true},{\"text\":\"edit\",\"url\":\"/development/masterfiles/users/$col0$/edit\",\"col0\":\"id\",\"icon\":\"edit\",\"title\":\"Edit\"},{\"text\":\"sep01\",\"is_separator\":true},{\"text\":\"delete\",\"url\":\"/development/masterfiles/users/$col0$\",\"col0\":\"id\",\"prompt\":\"Are you sure?\",\"method\":\"delete\",\"icon\":\"delete\",\"popup\":true}]"
+    assert_equal expected, actions
+  end
+
   def test_popup_and_loading_window_action
     DB.array_expect(BASIC_DATA)
     additions = { actions:
                   [{url: '/development/masterfiles/users/$:id$', text: 'view', icon: 'view-show', title: 'View', popup: true, loading_window: true}]
     }
-    assert_raises(ArgumentError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: { id: '1' }) }
+    assert_raises(ArgumentError) { Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' }) }
 
     ok_additions = { actions:
                   [{url: '/development/masterfiles/users/$:id$', text: 'view', icon: 'view-show', title: 'View', loading_window: true}]
     }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(ok_additions), params: { id: '1' })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(ok_additions), params: { id: '1' })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -383,7 +431,7 @@ class ListGridDataTest < Minitest::Test
   def test_calculated_columns
     DB.array_expect(BASIC_DATA)
     additions = { calculated_columns: [{ name: 'colnew', caption: 'ColCap', date_type: :number, format: :delimited_1000, expression: 'amount * id', position: 2 }] }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions), params: { id: '1' })
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions), params: { id: '1' })
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -401,7 +449,7 @@ class ListGridDataTest < Minitest::Test
                                                                                       'active' => { editor: :select, values: ['Yes', 'No'] },
                                                                                       'created_at' => { editor: :search_select, values: ['Now', 'Then', 'Soon'] },
                                                                                       'id' => { editor: :search_select, lookup_url: '/path/$:id$' } } } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions))
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions))
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       rows = data.list_rows
@@ -444,7 +492,7 @@ class ListGridDataTest < Minitest::Test
   def test_edit_select_rule
     DB.array_expect(BASIC_DATA)
     additions = { edit_rules: { url: '/path/to/$:id$/inline_save', editable_fields: { 'active' => { editor: :select, width: 350, value_sql: "SELECT t.* from (VALUES ('Yes'), ('No')) t" } } } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions))
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions))
     rows = nil
     data.stub(:load_report_def, BASIC_DM_REPORT) do
       data.stub(:select_editor_values, ['Yes', 'No']) do
@@ -462,7 +510,7 @@ class ListGridDataTest < Minitest::Test
   def test_edit_select_rule_ok
     DB.array_expect(BASIC_DATA)
     additions = { edit_rules: { url: '/path/to/$:id$/inline_save', editable_fields: { 'active' => { editor: :select, misspelled_values: [] } } } }
-    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, config_loader: loader_extended(additions))
+    data = Crossbeams::DataGrid::ListGridData.new(id: 'agrid', root_path: '/a/b/c', deny_access: ALLOW_ACCESS, has_permission: HAS_PERMISSION, client_rule_check: FALSE_CLIENT_RULE, config_loader: loader_extended(additions))
     assert_raises(ArgumentError) do
       data.stub(:load_report_def, BASIC_DM_REPORT) do
         rows = data.list_rows